requirements.txt
# Snowflake connectors and data processing
# Using patched version to fix SQL injection vulnerability in write_pandas
snowflake-connector-python>=3.13.1
snowflake-snowpark-python>=1.11.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Streamlit dashboard
streamlit>=1.29.0
plotly>=5.18.0

# Additional utilities
python-dotenv>=1.0.0
